+++ CMakeLists.txt
@@ -52,6 +52,7 @@ set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} ${TORCH_CXX_FLAGS}")
 
 add_library(lantern SHARED 
     src/lantern.cpp
+    src/tensor.cpp
 )
 add_library(lantern::library ALIAS lantern)
